MOD: Fix typos, renamed __utils to utils
